Use rawtxindex to fetch txs in REST server
Used for fetching the prevout txs of inputs. Not used by electrs internally elsewhere,
for the /tx/:txid endpoint or for the Electrum server, which still fetches via the
block height index and bitcoind.

This could be further optimized by batching reads, especially for the
prevout txs of the txs in a block.