Store block metadata in database
So that we don't have to fetch and parse the entire block to get it.

The metadata includes the block size, the number of transactions,
and the sum of transactions weights.

The `GET /blocks` and `GET /block/:hash` endpoints use the new metadata
and no longer have to fetch blocks from bitcoind.

The only remaining endpoint that talks to bitcoind is GET /block/:hash/txs.